Add: finish the creation of HttpMethodHandler and HttpErrorHandler to extract this logic from ServerEngine
